requirements.txt
streamlit>=1.38.0
openai>=1.40.0
langchain-openai>=0.1.17
chromadb>=0.5.5
pypdf>=4.3.1
pillow>=10.4.0
